backend/requirements.txt
# FastAPI框架和Web服务器
fastapi==0.109.2
uvicorn[standard]==0.27.0
python-multipart==0.0.7
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# 数据库
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.2.4

# 安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# HTTP客户端
httpx==0.26.0
aiohttp==3.9.3

# AI和ML
openai==1.13.3
anthropic==0.8.1
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0

# 文本处理
python-docx==1.1.0
PyPDF2==3.0.1
markdown==3.5.2
python-pptx==1.0.1
openpyxl==3.1.2
beautifulsoup4==4.12.3

# 日志和监控
loguru==0.7.2
prometheus-fastapi-instrumentator==6.1.0

# 缓存
redis==5.0.1

# 工具
tqdm==4.66.2
pillow==10.2.0
python-slugify==8.0.1
pytz==2023.3.post1
tenacity==8.2.3

# 测试
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.26.0